[refactoring] stp4: 패키지 분리
