+++ src/main/kotlin/lotto/lotto/LottoNumbers.kt
@@ -1,4 +1,4 @@
-package lotto_auto.lotto
+package lotto.lotto
 
 data class LottoNumbers(
     val lottoNumbers: List<Int>
