+++ src/main/kotlin/lotto/lotto/WinningLotto.kt
@@ -1,4 +1,4 @@
-package lotto_auto.lotto
+package lotto.lotto
 
 data class WinningLotto(
     val winningNumbers: LottoNumbers,
