+++ src/test/kotlin/lotto/LottoAutoTest.kt
@@ -1,8 +1,10 @@
-package lotto_auto
+package lotto
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
-import lotto_auto.lotto.*
+import lotto.lotto.*
+import lotto.lotto_auto.LottoAuto
+import lotto.lotto.LottoPrize
 
 class LottoAutoTest : StringSpec({
     "구매한 로또 총 당첨 금액이 맞는지" {
